add partial test
