+++ lib/ui/exporter/export_order_loader.dart
@@ -6,7 +6,7 @@ import 'package:possystem/models/objects/order_object.dart';
 import 'package:possystem/settings/currency_setting.dart';
 import 'package:possystem/translator.dart';
 
-class ExportOrderLoader extends StatefulWidget {
+class ExportOrderLoader extends StatelessWidget {
   final ValueNotifier<DateTimeRange> notifier;
 
   final Widget Function(OrderObject) formatOrder;
@@ -18,57 +18,25 @@ class ExportOrderLoader extends StatefulWidget {
   }) : super(key: key);
 
   @override
-  State<ExportOrderLoader> createState() => ExportOrderLoaderState();
-
-  static String formatCreatedAt(OrderObject order) {
-    return '${DateFormat.MMMd(S.localeName).format(order.createdAt)} ${DateFormat.Hms(S.localeName).format(order.createdAt)}';
-  }
-
-  static String formatHeader(OrderObject order) {
-    return [
-      '${order.totalCount} 份餐點',
-      '共 ${order.totalPrice.toCurrency()} 元',
-    ].join(MetaBlock.string);
-  }
-}
-
-class ExportOrderLoaderState extends State<ExportOrderLoader> {
-  final loaderKey = OrderLoader.createKey();
-
-  @override
   Widget build(BuildContext context) {
     return OrderLoader(
       calculateMemory: true,
-      loaderKey: loaderKey,
-      ranger: () => widget.notifier.value,
+      ranger: notifier,
       builder: _buildOrder,
     );
   }
 
-  @override
-  void initState() {
-    super.initState();
-    widget.notifier.addListener(_onRangeChanged);
-  }
-
-  @override
-  void dispose() {
-    widget.notifier.removeListener(_onRangeChanged);
-    super.dispose();
-  }
-
-  void _onRangeChanged() {
-    loaderKey.currentState?.reset();
-  }
-
-  Widget _buildOrder(OrderObject order) {
+  Widget _buildOrder(BuildContext context, OrderObject order) {
     return ListTile(
       leading: Padding(
         padding: const EdgeInsets.only(top: 4.0),
         child: Text(DateFormat.Hm(S.localeName).format(order.createdAt)),
       ),
-      title: Text(ExportOrderLoader.formatCreatedAt(order)),
-      subtitle: Text(ExportOrderLoader.formatHeader(order)),
+      title: Text(DateFormat('M月d日 HH:mm:ss\n').format(order.createdAt)),
+      subtitle: Text([
+        '${order.totalCount} 份餐點',
+        '共 ${order.totalPrice.toCurrency()} 元',
+      ].join(MetaBlock.string)),
       trailing: TextButton(
         onPressed: () {
           showDialog(
@@ -77,7 +45,7 @@ class ExportOrderLoaderState extends State<ExportOrderLoader> {
               return SimpleDialog(title: const Text('訂單細節'), children: [
                 Padding(
                   padding: const EdgeInsets.all(8.0),
-                  child: widget.formatOrder(order),
+                  child: formatOrder(order),
                 ),
               ]);
             },
